Resolve conflicts during dedupe of issues
This moves the logic of checking if an issue is really a PR to the import time
and removes the need for deduping after the scrape. The existing DB level dedupe
was removing issues that were not duplicate PR as there happened to be issues with
the same id as some PRs across repos (and even in the same repo. This was apparent
with fresh repos where it is common to have low issue and PR numbers.